Revamped the entire code. Redesigned the application - Now application does not match a buyer to a seller but rather shows a list of sellers to the buyer.
Also, MapMyIndia API is used to fetch the distance between different 
hostels and output it to the user in form of a table.

All the data storage and retrieval is done using JSON Library and the 
data is stored as .json file.

I will properly comment the code again in another commit if required.